Extract rate limiting semaphore to its own source file
